Move buttons between annotation list and actions
